major updates to user registration
put in a registration controller in the apps/controller directory.
Controller (when POST) is called, will validate the new user
registration and then add the user to the MySQL database.
Separation of concerns regarding the MVC (register user -> controller,
adding the user to the db -> model, database -> utilities/backend,
registration.html page -> view)